feat: util함수 재구성 및 추가, basic import 수정
