+++ packages/assignment-5/src/refactoring/hooks/utils/newId.ts
@@ -1,0 +1,1 @@
+export const createNewId = () => crypto.randomUUID();
